requirements.txt
Django==1.11
django-braces==1.11.0
django-oauth-toolkit==0.12.0
djangorestframework==3.6.2
oauthlib==2.0.1
psycopg2==2.7.1
pytz==2017.2
six==1.10.0
